+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - DELIBE-23/DELIBE-206: Complete the upgrade step 2100.
   [aduchene]
+- DELIBE-24: Fix an issue where a decisions manager could not remove items from a meeting.
+  [aduchene]
 
 2.1.0 (2025-02-28)
 ------------------